requirements.txt
# LostFuzzer v2.0 Python Dependencies
# ====================================

# Core
pyyaml>=6.0
rich>=13.0
typer>=0.9.0
python-dotenv>=1.0.0

# HTTP & Async
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0

# Parsing & Analysis
beautifulsoup4>=4.12.0
lxml>=4.9.0
jsbeautifier>=1.14.0
regex>=2023.0

# Reporting
jinja2>=3.1.0
markdown>=3.5.0

# Notifications
slack-sdk>=3.23.0
discord-webhook>=1.3.0
python-telegram-bot>=20.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
tabulate>=0.9.0
